organized factory folder
