+++ factory/src/error.rs
@@ -6,16 +6,12 @@ use cosmwasm_std::StdError;
 pub enum ContractError {
     #[error("{0}")]
     Std(#[from] StdError),
-
     
     #[error("Unauthorized")]
     Unauthorized {},
 
     #[error("InsufficientFunds")]
     InsufficientFunds {},
-
-    #[error("This user already claimed for this day")]
-    AlreadyClaimed {},
 
     #[error("Wrong configuration")]
     WrongConfiguration {},
@@ -33,5 +29,4 @@ pub enum ContractError {
     UnknownReplyId {
         id: u64,
     }
-
 }
