+++ factory/src/execute.rs
@@ -1,26 +1,22 @@
 use std::env;
-
-use crate::asset::AssetInfo;
 use crate::error::ContractError;
-use crate::msg::{CreatePoolReplyMsg, ExecuteMsg, MigrateMsg, TokenInfo, TokenInstantiateMsg};
-use crate::pair::{CreatePool, FeeInfo, ThresholdPayout};
+use crate::msg::{ ExecuteMsg, MigrateMsg, TokenInfo, TokenInstantiateMsg};
+use crate::pair::{CreatePool,};
 use crate::state::{
-    CommitInfo, CreationState, CreationStatus, FactoryInstantiate, COMMIT, CONFIG, CREATION_STATES,
-    NEXT_POOL_ID, POOLS_BY_ID, TEMPCREATOR, TEMPNFTADDR, TEMPPAIRINFO, TEMPPOOLID, TEMPTOKENADDR,
+    CreationState, CreationStatus, FactoryInstantiate, CONFIG, CREATION_STATES,
+    NEXT_POOL_ID, TEMPCREATOR, TEMPPAIRINFO, TEMPPOOLID,
 };
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::entry_point;
 use cosmwasm_std::{
-    to_json_binary, Addr, CosmosMsg, Deps, DepsMut, Empty, Env, MessageInfo, Reply, Response,
-    StdError, StdResult, Storage, SubMsg, SubMsgResponse, SubMsgResult, Uint128, WasmMsg,
+    to_json_binary, Deps, DepsMut, Env, MessageInfo,Response,
+    StdError, StdResult, SubMsg, Uint128, WasmMsg,
 };
-use cw20::{Cw20ExecuteMsg, MinterResponse};
-use cw721_base::msg::InstantiateMsg as Cw721InstantiateMsg;
-use cw721_base::Action;
+use cw20::{MinterResponse};
 
 const CONTRACT_NAME: &str = "bluechip_factory";
 const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
-const BURN_ADDRESS: &str = "cosmos1qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqnrql8a";
+pub const BURN_ADDRESS: &str = "cosmos1qqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqqnrql8a";
 pub const SET_TOKENS: u64 = 1;
 pub const MINT_CREATE_POOL: u64 = 2;
 pub const FINALIZE_POOL: u64 = 3;
@@ -149,7 +145,7 @@ fn execute_create(
         retry_count: 0,
     };
     CREATION_STATES.save(deps.storage, pool_id, &creation_state)?;
-    //triggers reply function when things go well.
+    //triggers reply function in reply.rs when things go well.
     let sub_msg = vec![SubMsg::reply_on_success(msg, SET_TOKENS)];
 
     Ok(Response::new()
@@ -159,351 +155,3 @@ fn execute_create(
         .add_submessages(sub_msg))
 }
 
-#[entry_point]
-//called by execute create.
-pub fn reply(deps: DepsMut, env: Env, msg: Reply) -> Result<Response, ContractError> {
-    match msg.id {
-        SET_TOKENS => set_tokens(deps, env, msg),
-        MINT_CREATE_POOL => mint_create_pool(deps, env, msg),
-        FINALIZE_POOL => finalize_pool(deps, env, msg),
-        CLEANUP_TOKEN_REPLY_ID => handle_cleanup_reply(deps, env, msg),
-        CLEANUP_NFT_REPLY_ID => handle_cleanup_reply(deps, env, msg),
-        _ => Err(ContractError::UnknownReplyId { id: msg.id }),
-    }
-}
-
-fn set_tokens(deps: DepsMut, env: Env, msg: Reply) -> Result<Response, ContractError> {
-    let pool_id = TEMPPOOLID.load(deps.storage)?;
-    let mut creation_state = CREATION_STATES.load(deps.storage, pool_id)?;
-    let factory_addr = env.contract.address.clone();
-    match msg.result {
-        SubMsgResult::Ok(result) => {
-            // extract token address from reply - helper below.
-            let token_address = extract_contract_address(&result)?;
-
-            //update creation state to catch transaction failures.
-            creation_state.token_address = Some(token_address.clone());
-            creation_state.status = CreationStatus::TokenCreated;
-            CREATION_STATES.save(deps.storage, pool_id, &creation_state)?;
-
-            // Save to temp storage for next step
-            TEMPTOKENADDR.save(deps.storage, &token_address)?;
-
-            // Create NFT instantiate message
-            let config = CONFIG.load(deps.storage)?;
-            let nft_instantiate_msg = to_json_binary(&Cw721InstantiateMsg {
-                name: "AMM LP Positions".to_string(),
-                symbol: "AMM-LP".to_string(),
-                minter: env.contract.address.to_string(),
-            })?;
-
-            let nft_msg = WasmMsg::Instantiate {
-                code_id: config.position_nft_id,
-                msg: nft_instantiate_msg,
-                //no initial NFTs since no positions have been created yet
-                funds: vec![],
-                //set factory as admin
-                admin: Some(factory_addr.to_string()),
-                label: format!("AMM-LP-NFT-{}", token_address),
-            };
-            //everything successful, move on to next step in creation
-            let sub_msg = SubMsg::reply_on_success(nft_msg, MINT_CREATE_POOL);
-
-            Ok(Response::new()
-                .add_attribute("action", "token_created_successfully")
-                .add_attribute("token_address", token_address)
-                .add_attribute("pool_id", pool_id.to_string())
-                .add_submessage(sub_msg))
-        }
-        SubMsgResult::Err(err) => {
-            // failure, mark as clean up and burn. Check helpers below.
-            creation_state.status = CreationStatus::Failed;
-            CREATION_STATES.save(deps.storage, pool_id, &creation_state)?;
-            cleanup_temp_state(deps.storage)?;
-
-            Err(ContractError::TokenCreationFailed {
-                pool_id,
-                reason: err,
-            })
-        }
-    }
-}
-
-fn mint_create_pool(deps: DepsMut, env: Env, msg: Reply) -> Result<Response, ContractError> {
-    let pool_id = TEMPPOOLID.load(deps.storage)?;
-    let mut creation_state = CREATION_STATES.load(deps.storage, pool_id)?;
-    let factory_addr = env.contract.address.clone();
-    match msg.result {
-        SubMsgResult::Ok(result) => {
-            let nft_address = extract_contract_address(&result)?;
-            //set creation state for tracking and potential errors
-            creation_state.nft_address = Some(nft_address.clone());
-            creation_state.status = CreationStatus::NftCreated;
-            CREATION_STATES.save(deps.storage, pool_id, &creation_state)?;
-
-            TEMPNFTADDR.save(deps.storage, &nft_address)?;
-
-            let config = CONFIG.load(deps.storage)?;
-            let temp_pool_info = TEMPPAIRINFO.load(deps.storage)?;
-            let temp_creator = TEMPCREATOR.load(deps.storage)?;
-            let token_address = TEMPTOKENADDR.load(deps.storage)?;
-
-            //set threshold payouts for threshold crossing
-            let threshold_payout = ThresholdPayout {
-                creator_amount: Uint128::new(325_000_000_000),
-                bluechip_amount: Uint128::new(25_000_000_000),
-                pool_amount: Uint128::new(350_000_000_000),
-                commit_amount: Uint128::new(500_000_000_000),
-            };
-            let threshold_binary = to_json_binary(&threshold_payout)?;
-
-            let mut updated_asset_infos = temp_pool_info.asset_infos;
-            for asset_info in updated_asset_infos.iter_mut() {
-                if let AssetInfo::Token { contract_addr } = asset_info {
-                    if contract_addr.as_str() == "WILL_BE_CREATED_BY_FACTORY" {
-                        *contract_addr = token_address.clone();
-                    }
-                }
-            }
-
-            // Instantiate the pool with final values
-            let pool_msg = WasmMsg::Instantiate {
-                code_id: config.pair_id,
-                msg: to_json_binary(&CreatePoolReplyMsg {
-                    pool_id,
-                    asset_infos: updated_asset_infos,
-                    factory_addr: env.contract.address,
-                    token_code_id: config.token_id,
-                    threshold_payout: Some(threshold_binary),
-                    fee_info: FeeInfo {
-                        bluechip_address: config.bluechip_address.clone(),
-                        creator_address: temp_creator.clone(),
-                        bluechip_fee: config.bluechip_fee,
-                        creator_fee: config.creator_fee,
-                    },
-                    commit_amount_for_threshold: config.commit_amount_for_threshold,
-                    commit_limit_usd: config.commit_limit_usd,
-                    oracle_addr: config.oracle_addr.clone(),
-                    oracle_symbol: config.oracle_symbol.clone(),
-                    token_address: token_address,
-                    position_nft_address: nft_address.clone(),
-                })?,
-                funds: vec![],
-                admin: Some(factory_addr.to_string()),
-                label: "Pair".to_string(),
-            };
-            //goes well, move on to next step
-            let sub_msg: SubMsg = SubMsg::reply_on_success(pool_msg, FINALIZE_POOL);
-
-            Ok(Response::new()
-                .add_attribute("action", "nft_created_successfully")
-                .add_attribute("nft_address", nft_address)
-                .add_attribute("pool_id", pool_id.to_string())
-                .add_submessage(sub_msg))
-        }
-        //doesnt go well, find which state the pool is in and behave accordingly.
-        SubMsgResult::Err(err) => {
-            creation_state.status = CreationStatus::CleaningUp;
-            CREATION_STATES.save(deps.storage, pool_id, &creation_state)?;
-
-            let cleanup_msgs = create_cleanup_messages(&creation_state)?;
-
-            Ok(Response::new()
-                .add_submessages(cleanup_msgs)
-                .add_attribute("action", "nft_creation_failed_cleanup")
-                .add_attribute("pool_id", pool_id.to_string())
-                .add_attribute("error", err))
-        }
-    }
-}
-
-//take created pool, give it minting rights, sync to necessary contracts to mint.
-fn finalize_pool(deps: DepsMut, _env: Env, msg: Reply) -> Result<Response, ContractError> {
-    let pool_id = TEMPPOOLID.load(deps.storage)?;
-    let mut creation_state = CREATION_STATES.load(deps.storage, pool_id)?;
-    match msg.result {
-        SubMsgResult::Ok(result) => {
-            let pool_address = extract_contract_address(&result)?;
-
-            creation_state.pool_address = Some(pool_address.clone());
-            creation_state.status = CreationStatus::PoolCreated;
-            CREATION_STATES.save(deps.storage, pool_id, &creation_state)?;
-
-            let temp_creator = TEMPCREATOR.load(deps.storage)?;
-            let temp_token_address = TEMPTOKENADDR.load(deps.storage)?;
-            let temp_nft_address = TEMPNFTADDR.load(deps.storage)?;
-            // create commit parameters for commit transactions
-            let commit_info = CommitInfo {
-                pool_id,
-                creator: temp_creator.clone(),
-                token_addr: temp_token_address.clone(),
-                pool_addr: pool_address.clone(),
-            };
-            // save commit state to record future commit executions
-            COMMIT.save(deps.storage, &temp_creator.to_string(), &commit_info)?;
-            POOLS_BY_ID.save(deps.storage, pool_id, &commit_info)?;
-
-            // make pool cw20 and cw721 (nft) minter for threshold payout - compartmentalizes pools so they can run without factory.
-            //1 factory for all pools instead of 1 factory 1 pool
-            let ownership_msgs = create_ownership_transfer_messages(
-                &temp_token_address,
-                &temp_nft_address,
-                &pool_address,
-            )?;
-            //mark completed and clean all temp states since all values have all been updated and created.
-            creation_state.status = CreationStatus::Completed;
-            CREATION_STATES.save(deps.storage, pool_id, &creation_state)?;
-
-            cleanup_temp_state(deps.storage)?;
-
-            Ok(Response::new()
-                .add_messages(ownership_msgs)
-                .add_attribute("action", "pool_created_successfully")
-                .add_attribute("pool_address", pool_address)
-                .add_attribute("pool_id", pool_id.to_string()))
-        }
-        SubMsgResult::Err(err) => {
-            // Pool creation failed - cleanup everything
-            creation_state.status = CreationStatus::CleaningUp;
-            CREATION_STATES.save(deps.storage, pool_id, &creation_state)?;
-
-            let cleanup_msgs = create_cleanup_messages(&creation_state)?;
-
-            Ok(Response::new()
-                .add_submessages(cleanup_msgs)
-                .add_attribute("action", "pool_creation_failed_cleanup")
-                .add_attribute("pool_id", pool_id.to_string())
-                .add_attribute("error", err))
-        }
-    }
-}
-
-//helpers for reply function
-
-//clean and remove all temp information
-fn cleanup_temp_state(storage: &mut dyn Storage) -> Result<(), ContractError> {
-    TEMPPOOLID.remove(storage);
-    TEMPPAIRINFO.remove(storage);
-    TEMPCREATOR.remove(storage);
-    TEMPTOKENADDR.remove(storage);
-    TEMPNFTADDR.remove(storage);
-    Ok(())
-}
-
-//if partial transaction happens
-fn create_cleanup_messages(creation_state: &CreationState) -> Result<Vec<SubMsg>, ContractError> {
-    // Changed return type to Vec<SubMsg>
-    let mut messages = vec![];
-
-    // if token was created, disable it
-    if let Some(token_addr) = &creation_state.token_address {
-        let disable_token_msg = WasmMsg::Execute {
-            contract_addr: token_addr.to_string(),
-            msg: to_json_binary(&Cw20ExecuteMsg::UpdateMinter {
-                new_minter: None, // remove minter entirely
-            })?,
-            funds: vec![],
-        };
-
-        // create SubMsg that will trigger reply handler
-        let sub_msg: SubMsg = SubMsg::reply_on_error(disable_token_msg, CLEANUP_TOKEN_REPLY_ID);
-        messages.push(sub_msg);
-    }
-
-    // if NFT was created, disable it
-    if let Some(nft_addr) = &creation_state.nft_address {
-        let disable_nft_msg = WasmMsg::Execute {
-            contract_addr: nft_addr.to_string(),
-            msg: to_json_binary(&cw721_base::ExecuteMsg::<Empty, Empty>::UpdateOwnership(
-                Action::TransferOwnership {
-                    new_owner: BURN_ADDRESS.to_string(),
-                    expiry: None,
-                },
-            ))?,
-            funds: vec![],
-        };
-
-        // Create SubMsg that will trigger reply handler
-        let sub_msg: SubMsg = SubMsg::reply_on_error(disable_nft_msg, CLEANUP_NFT_REPLY_ID);
-        messages.push(sub_msg);
-    }
-
-    Ok(messages)
-}
-
-fn handle_cleanup_reply(deps: DepsMut, _env: Env, msg: Reply) -> Result<Response, ContractError> {
-    // Handle cleanup completion/failure
-    match msg.result {
-        SubMsgResult::Ok(_) => {
-            // Cleanup succeeded - remove creation state
-            if let Ok(pool_id) = TEMPPOOLID.load(deps.storage) {
-                CREATION_STATES.remove(deps.storage, pool_id);
-                cleanup_temp_state(deps.storage)?;
-            }
-
-            Ok(Response::new().add_attribute("action", "cleanup_completed"))
-        }
-        SubMsgResult::Err(err) => {
-            
-            if let Ok(pool_id) = TEMPPOOLID.load(deps.storage) {
-                if let Ok(mut state) = CREATION_STATES.load(deps.storage, pool_id) {
-                    state.status = CreationStatus::Failed;
-                    state.retry_count += 1;
-                    CREATION_STATES.save(deps.storage, pool_id, &state)?;
-                }
-            }
-
-            Ok(Response::new()
-                .add_attribute("action", "cleanup_failed")
-                .add_attribute("error", err))
-        }
-    }
-}
-
-//pull contract addresss - can be used for multiple types.
-fn extract_contract_address(result: &SubMsgResponse) -> Result<Addr, ContractError> {
-    result
-        .events
-        .iter()
-        .find(|event| event.ty == "instantiate")
-        .and_then(|event| {
-            event
-                .attributes
-                .iter()
-                .find(|attr| attr.key == "_contract_address")
-                .map(|attr| attr.value.clone())
-        })
-        .ok_or_else(|| ContractError::ContractAddressNotFound {})
-        .and_then(|addr_str| Ok(Addr::unchecked(addr_str)))
-}
-
-//give pool minter responsibilities
-fn create_ownership_transfer_messages(
-    token_addr: &Addr,
-    nft_addr: &Addr,
-    pool_addr: &Addr,
-) -> Result<Vec<CosmosMsg>, ContractError> {
-    Ok(vec![
-        WasmMsg::Execute {
-            contract_addr: token_addr.to_string(),
-            msg: to_json_binary(&Cw20ExecuteMsg::UpdateMinter {
-                //make pool minter of the tokens, 
-                new_minter: Some(pool_addr.to_string()),
-            })?,
-            funds: vec![],
-        }
-        .into(),
-        WasmMsg::Execute {
-            contract_addr: nft_addr.to_string(),
-            msg: to_json_binary(&cw721_base::ExecuteMsg::<Empty, Empty>::UpdateOwnership(
-                Action::TransferOwnership {
-                    //pool now own nft contract to mint nfts to liquidity providers
-                    new_owner: pool_addr.to_string(),
-                    expiry: None,
-                },
-            ))?,
-            funds: vec![],
-        }
-        .into(),
-    ])
-}
