+++ factory/src/query.rs
@@ -1,10 +1,65 @@
-use cosmwasm_std::{Deps, StdResult};
-
+use cosmwasm_schema::{cw_serde, QueryResponses};
+use cosmwasm_std::{to_json_binary, Addr, BankQuery, Deps, QuerierWrapper, QueryRequest, StdResult, Uint128, WasmQuery};
+use cw20::{BalanceResponse, Cw20QueryMsg, TokenInfoResponse};
+use crate::pair::PairInfo;
 use crate::msg::ConfigResponse;
 use crate::state::CONFIG;
+
+
+#[cw_serde]
+#[derive(QueryResponses)]
+pub enum QueryMsg {
+    // factory config
+    #[returns(ConfigResponse)]
+    Config {},
+    #[returns(PairInfo)]
+    Pair {},
+}
 
 
 pub fn query_config(deps: Deps) -> StdResult<ConfigResponse> {
     let config = CONFIG.load(deps.storage)?;
     Ok(ConfigResponse { config })
+}
+
+pub fn query_token_balance(
+    querier: &QuerierWrapper,
+    contract_addr: Addr,
+    account_addr: Addr,
+) -> StdResult<Uint128> {
+    // load balance from the token contract
+    let res: BalanceResponse = querier
+        .query(&QueryRequest::Wasm(WasmQuery::Smart {
+            contract_addr: String::from(contract_addr),
+            msg: to_json_binary(&Cw20QueryMsg::Balance {
+                address: String::from(account_addr),
+            })?,
+        }))
+        .unwrap_or_else(|_| BalanceResponse {
+            balance: Uint128::zero(),
+        });
+
+    Ok(res.balance)
+}
+
+
+pub fn query_token_symbol(querier: &QuerierWrapper, contract_addr: Addr) -> StdResult<String> {
+    let res: TokenInfoResponse = querier.query(&QueryRequest::Wasm(WasmQuery::Smart {
+        contract_addr: String::from(contract_addr),
+        msg: to_json_binary(&Cw20QueryMsg::TokenInfo {})?,
+    }))?;
+
+    Ok(res.symbol)
+}
+
+pub fn query_balance(
+    querier: &QuerierWrapper,
+    account_addr: Addr,
+    denom: String,
+) -> StdResult<Uint128> {
+    let balance: BalanceResponse = querier.query(&QueryRequest::Bank(BankQuery::Balance {
+        address: String::from(account_addr),
+        denom,
+    }))?;
+    Ok(balance.amount.amount)
 }
